+++ win/win32/winclisrv/GnollHackClient/GnollHackClient.Android/GnollHackService.cs
@@ -389,7 +389,7 @@ namespace GnollHackClient.Droid
             return RunGnollHack(
                 _gnollhackfilesdir,
                 "",
-                (UInt32)(clientGame.WizardMode ? RunGnollHackFlags.WizardMode : 0),
+                (ulong)(clientGame.WizardMode ? RunGnollHackFlags.WizardMode : 0),
                 0,
                 0,
                 clientGame.ClientCallback_InitWindows,
