+++ win/win32/winclisrv/GnollHackClient/GnollHackClient/IFmodService.cs
@@ -12,6 +12,6 @@ namespace GnollHackClient
         void StopTestSound();
         int PlayImmediateSound(int ghsound, string eventPath, int bankid, float eventVolume, float soundVolume, string[] parameterNames, float[] parameterValues, int arraysize, int sound_type, int play_group, uint dialogue_mid);
         int PlayMusic(int ghsound, string eventPath, int bankid, float eventVolume, float soundVolume);
-        int StopAllSounds(UInt32 flags, uint dialogue_mid);
+        int StopAllSounds(ulong flags, uint dialogue_mid);
     }
 }
